requirements.txt
newspaper3k
sklearn
numpy
nltk